fix: typo for lowercase events
