+++ lib/confBridgeConfUser.js
