+++ lib/confBridgeConference.js
