+++ lib/plugins_command_13/confBridgeConfUserMute.js
@@ -11,9 +11,9 @@ var action = require('../action');
  * @private
  * @final
  * @readOnly
- * @default [confBridgeConfUserUnmute]
+ * @default [confBridgeConfUserMute]
  */
-var IDLOG = '[confBridgeConfUserUnmute]';
+var IDLOG = '[confBridgeConfUserMute]';
 
 (function () {
 
@@ -39,21 +39,21 @@ var IDLOG = '[confBridgeConfUserUnmute]';
     var map = {};
 
     /**
-     * Command plugin to unmute a ConfBridge user.
+     * Command plugin to mute a ConfBridge user.
      *
      * Use it with _astproxy_ module as follow:
      *
-     *     astproxy.doCmd({ command: 'confBridgeConfUserUnmute', confId: '6666202', usernum: '2', confBridgeConfCode: '1234' }, function (res) {
+     *     astproxy.doCmd({ command: 'confBridgeConfUserMute', confId: '6666202', usernum: '2', confBridgeConfCode: '1234' }, function (res) {
      *         // some code
      *     });
      *
-     * @class confBridgeConfUserUnmute
+     * @class confBridgeConfUserMute
      * @static
      */
-    var confBridgeConfUserUnmute = {
+    var confBridgeConfUserMute = {
 
       /**
-       * Execute asterisk action to unmute a ConfBridge user.
+       * Execute asterisk action to mute a ConfBridge user.
        *
        * @method execute
        * @param {object}   am   Asterisk manager to send the action
@@ -65,13 +65,13 @@ var IDLOG = '[confBridgeConfUserUnmute]';
         try {
           // action for asterisk
           var act = {
-            Action: 'ConfBridgeUnmute',
+            Action: 'ConfbridgeMute',
             ConfBridge: args.confBridgeConfCode + args.confId,
             Usernum: args.usernum
           };
 
           // set the action identifier
-          act.ActionID = action.getActionId('confBridgeConfUserUnmute');
+          act.ActionID = action.getActionId('confBridgeConfUserMute');
 
           // add association ActionID-callback
           map[act.ActionID] = cb;
@@ -146,9 +146,9 @@ var IDLOG = '[confBridgeConfUserUnmute]';
     };
 
     // public interface
-    exports.data = confBridgeConfUserUnmute.data;
-    exports.execute = confBridgeConfUserUnmute.execute;
-    exports.setLogger = confBridgeConfUserUnmute.setLogger;
+    exports.data = confBridgeConfUserMute.data;
+    exports.execute = confBridgeConfUserMute.execute;
+    exports.setLogger = confBridgeConfUserMute.setLogger;
 
   } catch (err) {
     logger.error(IDLOG, err.stack);
