+++ lib/plugins_command_13/confBridgeConfUserUnmute.js
@@ -11,9 +11,9 @@ var action = require('../action');
  * @private
  * @final
  * @readOnly
- * @default [confBridgeConfUserMute]
+ * @default [confBridgeConfUserUnmute]
  */
-var IDLOG = '[confBridgeConfUserMute]';
+var IDLOG = '[confBridgeConfUserUnmute]';
 
 (function () {
 
@@ -39,21 +39,21 @@ var IDLOG = '[confBridgeConfUserMute]';
     var map = {};
 
     /**
-     * Command plugin to mute a ConfBridge user.
+     * Command plugin to unmute a ConfBridge user.
      *
      * Use it with _astproxy_ module as follow:
      *
-     *     astproxy.doCmd({ command: 'confBridgeConfUserMute', confId: '6666202', usernum: '2', confBridgeConfCode: '1234' }, function (res) {
+     *     astproxy.doCmd({ command: 'confBridgeConfUserUnmute', confId: '6666202', usernum: '2', confBridgeConfCode: '1234' }, function (res) {
      *         // some code
      *     });
      *
-     * @class confBridgeConfUserMute
+     * @class confBridgeConfUserUnmute
      * @static
      */
-    var confBridgeConfUserMute = {
+    var confBridgeConfUserUnmute = {
 
       /**
-       * Execute asterisk action to mute a ConfBridge user.
+       * Execute asterisk action to unmute a ConfBridge user.
        *
        * @method execute
        * @param {object}   am   Asterisk manager to send the action
@@ -65,13 +65,13 @@ var IDLOG = '[confBridgeConfUserMute]';
         try {
           // action for asterisk
           var act = {
-            Action: 'ConfBridgeMute',
+            Action: 'ConfbridgeUnmute',
             ConfBridge: args.confBridgeConfCode + args.confId,
             Usernum: args.usernum
           };
 
           // set the action identifier
-          act.ActionID = action.getActionId('confBridgeConfUserMute');
+          act.ActionID = action.getActionId('confBridgeConfUserUnmute');
 
           // add association ActionID-callback
           map[act.ActionID] = cb;
@@ -146,9 +146,9 @@ var IDLOG = '[confBridgeConfUserMute]';
     };
 
     // public interface
-    exports.data = confBridgeConfUserMute.data;
-    exports.execute = confBridgeConfUserMute.execute;
-    exports.setLogger = confBridgeConfUserMute.setLogger;
+    exports.data = confBridgeConfUserUnmute.data;
+    exports.execute = confBridgeConfUserUnmute.execute;
+    exports.setLogger = confBridgeConfUserUnmute.setLogger;
 
   } catch (err) {
     logger.error(IDLOG, err.stack);
