+++ lib/plugins_command_13/confBridgeList.js
@@ -75,7 +75,7 @@ var IDLOG = '[confBridgeList]';
         try {
           // action for asterisk
           var act = {
-            Action: 'ConfBridgeList',
+            Action: 'ConfbridgeList',
             Conference: args.confBridgeConfCode + args.confId
           };
 
@@ -104,7 +104,7 @@ var IDLOG = '[confBridgeList]';
       data: function (data) {
         try {
           // store new member object
-          if (data.event === 'ConfBridgeList') {
+          if (data.event === 'ConfbridgeList') {
             var obj = {
               conference: data.conference,
               userNumber: data.usernumber,
@@ -120,10 +120,10 @@ var IDLOG = '[confBridgeList]';
             };
             list[data.usernumber] = obj;
 
-          } else if (map[data.actionid] && data.event === 'ConfBridgeListComplete') {
+          } else if (map[data.actionid] && data.event === 'ConfbridgeListComplete') {
             map[data.actionid](null, list);
           }
-          if (data.event === 'ConfBridgeListComplete') {
+          if (data.event === 'ConfbridgeListComplete') {
             list = {};
             delete map[data.actionid];
           }
