+++ lib/plugins_command_13/listConfBridgeConf.js
@@ -105,7 +105,7 @@ var IDLOG = '[listConfBridgeConf]';
 
           // action for asterisk
           var act = {
-            Action: 'ConfBridgeList'
+            Action: 'ConfbridgeList'
           };
           if (args.confId) {
             act.Conference = CONFBRIDGE_CONF_CODE + args.confId;
@@ -138,7 +138,7 @@ var IDLOG = '[listConfBridgeConf]';
           // check callback and info presence and execute it
           if (map[data.actionid] &&
             (
-              (data.eventlist === 'Complete' && data.event === 'ConfBridgeListComplete') ||
+              (data.eventlist === 'Complete' && data.event === 'ConfbridgeListComplete') ||
               (data.response === 'Error' && data.message === 'No active conferences.')
             )) {
 
@@ -149,11 +149,11 @@ var IDLOG = '[listConfBridgeConf]';
           } else if (map[data.actionid] &&
             data.channel &&
             data.conference &&
-            data.usernumber &&
+            data.accountcode &&
             data.calleridnum &&
             data.calleridname &&
             data.muted &&
-            data.event === 'ConfBridgeList') {
+            data.event === 'ConfbridgeList') {
 
             // filter extensions of remote sites. If the involved number is of remote site
             // it starts with phone prefix of remote site. So it checks the first character
@@ -170,7 +170,7 @@ var IDLOG = '[listConfBridgeConf]';
 
             var extenOwner = data.conference.substring(CONFBRIDGE_CONF_CODE.length, data.conference.length);
             var userObj = {
-              id: data.usernumber,
+              id: data.accountcode,
               name: data.calleridname === '<no name>' ? undefined : data.calleridname,
               site: remoteSite,
               muted: data.muted.toLowerCase() === 'no' ? false : true,
