+++ lib/plugins_event_13/confbridgeend.js
@@ -54,7 +54,7 @@ var astProxy;
        */
       data: function (data) {
         try {
-          if (data.conference && data.event === 'ConfBridgeEnd') {
+          if (data.conference && data.event === 'ConfbridgeEnd') {
 
             logger.info(IDLOG, 'received event ' + data.event);
             var CONFBRIDGE_CONF_CODE = astProxy.proxyLogic.getConfBridgeConfCode();
