+++ lib/plugins_event_13/confbridgejoin.js
@@ -11,9 +11,9 @@
  * @private
  * @final
  * @readOnly
- * @default [confBridgeMute]
+ * @default [confbridgeJoin]
  */
-var IDLOG = '[confBridgeMute]';
+var IDLOG = '[confbridgeJoin]';
 
 /**
  * The asterisk proxy.
@@ -38,15 +38,15 @@ var astProxy;
 
   try {
     /**
-     * The plugin that handles the ConfBridgeMute event.
+     * The plugin that handles the confbridgeJoin event.
      *
-     * @class confBridgeMute
+     * @class confbridgeJoin
      * @static
      */
-    var confBridgeMute = {
+    var confbridgeJoin = {
       /**
        * It's called from _astproxy_ component for each
-       * ConfBridgeMute event received from the asterisk.
+       * confbridgeJoin event received from the asterisk.
        *
        * @method data
        * @param {object} data The asterisk event data
@@ -54,23 +54,24 @@ var astProxy;
        */
       data: function (data) {
         try {
-          if (data.status &&
-            data.conference &&
-            data.user &&
-            data.event === 'ConfBridgeMute') {
+          if (data.conference &&
+            data.calleridnum &&
+            data.calleridname &&
+            data.event === 'ConfbridgeJoin') {
 
             logger.info(IDLOG, 'received event ' + data.event);
 
             var CONFBRIDGE_CONF_CODE = astProxy.proxyLogic.getConfBridgeConfCode();
             var extOwnerId = data.conference.substring(CONFBRIDGE_CONF_CODE.length, data.conference.length);
-            astProxy.proxyLogic.evtConfBridgeUserMute({
-              mute: data.status === 'on' ? true : false,
+            astProxy.proxyLogic.evtAddConfBridgeUserConf({
+              name: data.calleridname,
               confId: extOwnerId,
-              userId: data.user
+              userId: data.calleridnum,
+              extenId: data.calleridnum
             });
 
           } else {
-            logger.warn(IDLOG, 'ConfBridgeMute event not recognized');
+            logger.warn(IDLOG, 'ConfbridgeJoin event not recognized');
           }
 
         } catch (err) {
@@ -122,9 +123,9 @@ var astProxy;
     };
 
     // public interface
-    exports.data = confBridgeMute.data;
-    exports.visit = confBridgeMute.visit;
-    exports.setLogger = confBridgeMute.setLogger;
+    exports.data = confbridgeJoin.data;
+    exports.visit = confbridgeJoin.visit;
+    exports.setLogger = confbridgeJoin.setLogger;
 
   } catch (err) {
     logger.error(IDLOG, err.stack);
