+++ lib/plugins_event_13/confbridgeleave.js
@@ -58,7 +58,7 @@ var astProxy;
             data.user &&
             data.calleridnum &&
             data.calleridname &&
-            data.event === 'ConfBridgeLeave') {
+            data.event === 'ConfbridgeLeave') {
 
             logger.info(IDLOG, 'received event ' + data.event);
