+++ lib/plugins_event_13/confbridgemute.js
@@ -11,9 +11,9 @@
  * @private
  * @final
  * @readOnly
- * @default [confbridgeJoin]
+ * @default [confBridgeMute]
  */
-var IDLOG = '[confbridgeJoin]';
+var IDLOG = '[confBridgeMute]';
 
 /**
  * The asterisk proxy.
@@ -38,15 +38,15 @@ var astProxy;
 
   try {
     /**
-     * The plugin that handles the confbridgeJoin event.
+     * The plugin that handles the ConfBridgeMute event.
      *
-     * @class confbridgeJoin
+     * @class confBridgeMute
      * @static
      */
-    var confbridgeJoin = {
+    var confBridgeMute = {
       /**
        * It's called from _astproxy_ component for each
-       * confbridgeJoin event received from the asterisk.
+       * ConfBridgeMute event received from the asterisk.
        *
        * @method data
        * @param {object} data The asterisk event data
@@ -54,24 +54,23 @@ var astProxy;
        */
       data: function (data) {
         try {
-          if (data.conference &&
-            data.calleridnum &&
-            data.calleridname &&
-            data.event === 'ConfbridgeJoin') {
+          if (data.status &&
+            data.conference &&
+            data.user &&
+            data.event === 'ConfbridgeMute') {
 
             logger.info(IDLOG, 'received event ' + data.event);
 
-            var CONFBRIDGE_CONF_CODE = astProxy.proxyLogic.getConfbridgeConfCode();
+            var CONFBRIDGE_CONF_CODE = astProxy.proxyLogic.getConfBridgeConfCode();
             var extOwnerId = data.conference.substring(CONFBRIDGE_CONF_CODE.length, data.conference.length);
-            astProxy.proxyLogic.evtAddConfbridgeUserConf({
-              name: data.calleridname,
+            astProxy.proxyLogic.evtConfBridgeUserMute({
+              mute: data.status === 'on' ? true : false,
               confId: extOwnerId,
-              userId: data.calleridnum,
-              extenId: data.calleridnum
+              userId: data.user
             });
 
           } else {
-            logger.warn(IDLOG, 'ConfbridgeJoin event not recognized');
+            logger.warn(IDLOG, 'ConfBridgeMute event not recognized');
           }
 
         } catch (err) {
@@ -123,9 +122,9 @@ var astProxy;
     };
 
     // public interface
-    exports.data = confbridgeJoin.data;
-    exports.visit = confbridgeJoin.visit;
-    exports.setLogger = confbridgeJoin.setLogger;
+    exports.data = confBridgeMute.data;
+    exports.visit = confBridgeMute.visit;
+    exports.setLogger = confBridgeMute.setLogger;
 
   } catch (err) {
     logger.error(IDLOG, err.stack);
