+++ lib/proxy_logic_13/proxy_logic_13.js
@@ -28,8 +28,8 @@ var EventEmitter = require('events').EventEmitter;
 var ParkedCaller = require('../parkedCaller').ParkedCaller;
 var Conversation = require('../conversation').Conversation;
 var utilChannel13 = require('./util_channel_13');
-var ConfBridgeConfUser = require('../meetmeConfUser').ConfBridgeConfUser;
-var ConfBridgeConference = require('../meetmeConference').ConfBridgeConference;
+var ConfBridgeConfUser = require('../confBridgeConfUser').ConfBridgeConfUser;
+var ConfBridgeConference = require('../confBridgeConference').ConfBridgeConference;
 var RECORDING_STATUS = require('../conversation').RECORDING_STATUS;
 var TrunkConversation = require('../trunkConversation').TrunkConversation;
 var EXTEN_STATUS_ENUM = require('../extension').EXTEN_STATUS_ENUM;
